update rest streaming logic
