+++ google/api_core/rest_streaming_async.py
@@ -23,10 +23,10 @@ import requests
 import google.protobuf.message
 from google.protobuf.json_format import Parse
 import aiohttp
+from google.api_core._rest_streaming_base import BaseResponseIterator
 
-
-class ResponseAsyncIterator:
-    """Iterator over REST API responses.
+class ResponseAsyncIterator(BaseResponseIterator):
+    """Iterator over REST API Asynchronous responses.
 
     Args:
         response (requests.Response): An API response object.
@@ -43,88 +43,27 @@ class ResponseAsyncIterator:
         response_message_cls: Union[proto.Message, google.protobuf.message.Message],
     ):
         self._response = response
-        self._response_message_cls = response_message_cls
-        # Inner iterator over HTTP response's content.
-        # self._response_itr = self._response.iter_content(decode_unicode=True)
-        self._response_itr = self._response.content.iter_any()
-        # Contains a list of JSON responses ready to be sent to user.
-        self._ready_objs: Deque[str] = deque()
-        # Current JSON response being built.
-        self._obj = ""
-        # Keeps track of the nesting level within a JSON object.
-        self._level = 0
-        # Keeps track whether HTTP response is currently sending values
-        # inside of a string value.
-        self._in_string = False
-        # Whether an escape symbol "\" was encountered.
-        self._escape_next = False
-        self._data = None
-
-    def _process_chunk(self, chunk: str):
-        if self._level == 0:
-            if chunk[0] != "[":
-                raise ValueError(
-                    "Can only parse array of JSON objects, instead got %s" % chunk
-                )
-        for char in chunk:
-            if char == "{":
-                if self._level == 1:
-                    # Level 1 corresponds to the outermost JSON object
-                    # (i.e. the one we care about).
-                    self._obj = ""
-                if not self._in_string:
-                    self._level += 1
-                self._obj += char
-            elif char == "}":
-                self._obj += char
-                if not self._in_string:
-                    self._level -= 1
-                if not self._in_string and self._level == 1:
-                    self._ready_objs.append(self._obj)
-            elif char == '"':
-                # Helps to deal with an escaped quotes inside of a string.
-                if not self._escape_next:
-                    self._in_string = not self._in_string
-                self._obj += char
-            elif char in string.whitespace:
-                if self._in_string:
-                    self._obj += char
-            elif char == "[":
-                if self._level == 0:
-                    self._level += 1
-                else:
-                    self._obj += char
-            elif char == "]":
-                if self._level == 1:
-                    self._level -= 1
-                else:
-                    self._obj += char
-            else:
-                self._obj += char
-            self._escape_next = not self._escape_next if char == "\\" else False
+        self._response_itr = self._response.content.iter_chunked(1024)
+        super(ResponseAsyncIterator, self).__init__(response_message_cls=response_message_cls)
     
     async def __anext__(self):
-        if not self._data:
-            self._data = await self._response.read()
-            self._process_chunk(self._data.decode('utf-8'))
-        
-        if not self._ready_objs:
-            if self._level > 0:
-                raise ValueError("Unfinished stream: %s" % self._obj)
-            raise StopAsyncIteration
-        
+        while not self._ready_objs:
+            try:
+                chunk = await self._response_itr.__anext__()
+                chunk = chunk.decode('utf-8')
+                self._process_chunk(chunk)
+            except StopAsyncIteration as e:
+                if self._level > 0:
+                    raise ValueError("Unfinished stream: %s" % self._obj)
+                raise e
         return self._grab()
-
-    def _grab(self):
-        # Add extra quotes to make json.loads happy.
-        if issubclass(self._response_message_cls, proto.Message):
-            return self._response_message_cls.from_json(self._ready_objs.popleft())
-        elif issubclass(self._response_message_cls, google.protobuf.message.Message):
-            return Parse(self._ready_objs.popleft(), self._response_message_cls())
-        else:
-            raise ValueError(
-                "Response message class must be a subclass of proto.Message or google.protobuf.message.Message."
-            )
     
     def __aiter__(self):
         return self
+
+    
+    async def __aexit__(self, exc_type, exc, tb):
+        await self._response.release()
+
+    async def cancel(self):
+        await self._response.close()
